register

+++ pills/src/main/java/org/pablog/pills/web/ProductController.java
@@ -1,26 +1,23 @@
 package org.pablog.pills.web;
 import java.io.UnsupportedEncodingException;
-import java.util.List;
+import java.util.ArrayList;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
 import org.bson.types.ObjectId;
 import org.pablog.pills.domain.Product;
+import org.pablog.pills.domain.User;
 import org.pablog.pills.repositories.ProductRepository;
+import org.pablog.pills.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.util.UriUtils;
 import org.springframework.web.util.WebUtils;
 
@@ -28,17 +25,18 @@ import org.springframework.web.util.WebUtils;
 @Controller
 public class ProductController {
 
-	@Autowired
-    ProductRepository productRepository;
+	@Autowired ProductRepository productRepository;
+	@Autowired UserRepository userRepository;
 
 	@RequestMapping(method = RequestMethod.POST, produces = "text/html")
-    public String create(@Valid Product product, BindingResult bindingResult, Model uiModel, HttpServletRequest httpServletRequest) {
+    public String create(@ActiveUser User activeUser, @Valid Product product, BindingResult bindingResult, Model uiModel, HttpServletRequest httpServletRequest) {
         if (bindingResult.hasErrors()) {
             populateEditForm(uiModel, product);
             return "products/create";
         }
         uiModel.asMap().clear();
-        productRepository.save(product);
+        activeUser.getProducts().add(product);
+        userRepository.save(activeUser);
         return "redirect:/products/" + encodeUrlPathSegment(product.getId().toString(), httpServletRequest);
     }
 
@@ -56,16 +54,16 @@ public class ProductController {
     }
 
 	@RequestMapping(produces = "text/html")
-    public String list(@RequestParam(value = "page", required = false) Integer page, @RequestParam(value = "size", required = false) Integer size, Model uiModel) {
-        if (page != null || size != null) {
-            int sizeNo = size == null ? 10 : size.intValue();
-            final int firstResult = page == null ? 0 : (page.intValue() - 1) * sizeNo;
-            uiModel.addAttribute("products", productRepository.findAll(new org.springframework.data.domain.PageRequest(firstResult / sizeNo, sizeNo)).getContent());
-            float nrOfPages = (float) productRepository.count() / sizeNo;
-            uiModel.addAttribute("maxPages", (int) ((nrOfPages > (int) nrOfPages || nrOfPages == 0.0) ? nrOfPages + 1 : nrOfPages));
-        } else {
-            uiModel.addAttribute("products", productRepository.findAll());
-        }
+    public String list(@ActiveUser User activeUser, @RequestParam(value = "page", required = false) Integer page, @RequestParam(value = "size", required = false) Integer size, Model uiModel) {
+//        if (page != null || size != null) {
+//            int sizeNo = size == null ? 15 : size.intValue();
+//            final int firstResult = page == null ? 0 : (page.intValue() - 1) * sizeNo;
+//            uiModel.addAttribute("products", activeUser.getProducts(new PageRequest(firstResult / sizeNo, sizeNo)));
+//            float nrOfPages = (float) activeUser.getProducts().size() / sizeNo;
+//            uiModel.addAttribute("maxPages", (int) ((nrOfPages > (int) nrOfPages || nrOfPages == 0.0) ? nrOfPages + 1 : nrOfPages));
+//        } else {
+            uiModel.addAttribute("products", activeUser.getProducts());
+//        }
         return "products/list";
     }
 
@@ -87,9 +85,13 @@ public class ProductController {
     }
 
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE, produces = "text/html")
-    public String delete(@PathVariable("id") ObjectId id, @RequestParam(value = "page", required = false) Integer page, @RequestParam(value = "size", required = false) Integer size, Model uiModel) {
-        Product product = productRepository.findById(id);
-        productRepository.delete(product);
+    public String delete(@ActiveUser User activeUser, @PathVariable("id") ObjectId id, @RequestParam(value = "page", required = false) Integer page, @RequestParam(value = "size", required = false) Integer size, Model uiModel) {
+		Product product = productRepository.findById(id);
+		if(activeUser.getProducts().contains(product)) {
+			activeUser.getProducts().remove(product);
+			userRepository.save(activeUser);
+			productRepository.delete(product);
+		}
         uiModel.asMap().clear();
         uiModel.addAttribute("page", (page == null) ? "1" : page.toString());
         uiModel.addAttribute("size", (size == null) ? "10" : size.toString());
@@ -109,68 +111,5 @@ public class ProductController {
             pathSegment = UriUtils.encodePathSegment(pathSegment, enc);
         } catch (UnsupportedEncodingException uee) {}
         return pathSegment;
-    }
-
-	@RequestMapping(value = "/{id}", method = RequestMethod.GET, headers = "Accept=application/json")
-    @ResponseBody
-    public ResponseEntity<String> showJson(@PathVariable("id") ObjectId id) {
-        Product product = productRepository.findById(id);
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("Content-Type", "application/json; charset=utf-8");
-        if (product == null) {
-            return new ResponseEntity<String>(headers, HttpStatus.NOT_FOUND);
-        }
-        return new ResponseEntity<String>(product.toJson(), headers, HttpStatus.OK);
-    }
-
-	@RequestMapping(headers = "Accept=application/json")
-    @ResponseBody
-    public ResponseEntity<String> listJson() {
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("Content-Type", "application/json; charset=utf-8");
-        List<Product> result = productRepository.findAll();
-        return new ResponseEntity<String>(Product.toJsonArray(result), headers, HttpStatus.OK);
-    }
-
-	@RequestMapping(method = RequestMethod.POST, headers = "Accept=application/json")
-    public ResponseEntity<String> createFromJson(@RequestBody String json) {
-        Product product = Product.fromJsonToProduct(json);
-        productRepository.save(product);
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("Content-Type", "application/json");
-        return new ResponseEntity<String>(headers, HttpStatus.CREATED);
-    }
-
-	@RequestMapping(value = "/jsonArray", method = RequestMethod.POST, headers = "Accept=application/json")
-    public ResponseEntity<String> createFromJsonArray(@RequestBody String json) {
-        for (Product product: Product.fromJsonArrayToProducts(json)) {
-            productRepository.save(product);
-        }
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("Content-Type", "application/json");
-        return new ResponseEntity<String>(headers, HttpStatus.CREATED);
-    }
-
-	@RequestMapping(value = "/{id}", method = RequestMethod.PUT, headers = "Accept=application/json")
-    public ResponseEntity<String> updateFromJson(@RequestBody String json, @PathVariable("id") ObjectId id) {
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("Content-Type", "application/json");
-        Product product = Product.fromJsonToProduct(json);
-        if (productRepository.save(product) == null) {
-            return new ResponseEntity<String>(headers, HttpStatus.NOT_FOUND);
-        }
-        return new ResponseEntity<String>(headers, HttpStatus.OK);
-    }
-
-	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE, headers = "Accept=application/json")
-    public ResponseEntity<String> deleteFromJson(@PathVariable("id") ObjectId id) {
-        Product product = productRepository.findById(id);
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("Content-Type", "application/json");
-        if (product == null) {
-            return new ResponseEntity<String>(headers, HttpStatus.NOT_FOUND);
-        }
-        productRepository.delete(product);
-        return new ResponseEntity<String>(headers, HttpStatus.OK);
     }
 }
